change attendee schema and debug the app

+++ controllers/eventControllers.js
@@ -1,9 +1,11 @@
 const Event = require("../models/event");
 const Attendance = require("../models/attendance");
 const { catchAsyncError } = require("../utils/catchAsyncError");
-const { ErrorHandler } = require("../utils/errorHandler");
+const ErrorHandler = require("../utils/errorHandler");
 
 /**
+ * To create a new event
+ * It will create an event and its attendance
  * Request Body
  * eventName - String
  * description - String
@@ -21,7 +23,9 @@ exports.createEvent = catchAsyncError(async (req, res, next) => {
   let attendanceConfirmation;
   if (eventConfirmation) {
     attendanceConfirmation = await attendance.save();
-  } else throw new ErrorHandler("Could not create event", "500");
+  } else {
+    throw new ErrorHandler("Could not create event", "500");
+  }
 
   res.status(200).json({
     success: true,
@@ -32,6 +36,7 @@ exports.createEvent = catchAsyncError(async (req, res, next) => {
 
 /**
  * To get all the events and there details
+ * It will send back all the events available right now in events collection
  */
 exports.getAllEvents = catchAsyncError(async (req, res, next) => {
   const events = await Event.find();
@@ -42,7 +47,8 @@ exports.getAllEvents = catchAsyncError(async (req, res, next) => {
 });
 
 /**
- * To get an event's report
+ * To get an event's report by its id
+ *
  */
 exports.getEventReport = catchAsyncError(async (req, res, next) => {
   const eventID = req.params.id;
@@ -56,18 +62,22 @@ exports.getEventReport = catchAsyncError(async (req, res, next) => {
       event: { ...eventDetails._doc, attendance },
     });
   } else {
-    res.status(404).json({
-      success: false,
-      msg: "event does not exist",
-    });
+    throw new ErrorHandler("Event does not exist", 404);
   }
 });
 
 /**
- * To change event details
+ * To change event details by its id
+ * Request body can contain any of its subsets
+ * eventName - String
+ * description - String
+ * date - String (ISO string)
+ * venue - String
  */
 exports.changeEventDetails = catchAsyncError(async (req, res, next) => {
   const confirmation = await Event.updateOne({ _id: req.params.id }, req.body);
+  if (confirmation.matchedCount === 0)
+    throw new ErrorHandler("Event does not exist", 404);
   res.status(200).json({
     success: confirmation.acknowledged,
     modifiedCount: confirmation.modifiedCount,
@@ -75,10 +85,11 @@ exports.changeEventDetails = catchAsyncError(async (req, res, next) => {
 });
 
 /**
- * To delete an event
+ * To delete an event by its id
  */
 exports.deleteEvent = catchAsyncError(async (req, res, next) => {
   const deletedEvent = await Event.findByIdAndDelete(req.params.id);
+  if (!deletedEvent) throw new ErrorHandler("Event does not exist", 404);
   await Attendance.findByIdAndDelete(deletedEvent.attendance);
   res.status(200).json({
     success: true,
